Sentigram/requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.0.3
djangorestframework==3.15.1
joblib==1.3.2
nltk==3.8.1
numpy==1.26.4
regex==2023.12.25
scikit-learn==1.4.1.post1
scipy==1.12.0
sqlparse==0.4.4
threadpoolctl==3.4.0
tqdm==4.66.2
